Added some traveller type unit tests

+++ app/controllers/TravellerTypeController.java
@@ -184,12 +184,17 @@ public class TravellerTypeController {
         }  else {
             try {
                 TravellerType travellerType = TravellerType.find.byId(typeId);
-                destination.deleteTravellerType(travellerType);
-                destination.update();
+                if(destination.getUser().getUserid() == user.getUserid()) {
+                    destination.deleteTravellerType(travellerType);
+                    destination.update();
+                    return redirect(routes.TravellerTypeController.updateDestinationTravellerType(destId));
+                }
+                else{
+                    return unauthorized("Oops, you do not own this destination.");
+                }
             } catch (NumberFormatException e) {
                 return unauthorized("Oops, you do not have any traveller types to delete");
             }
         }
-        return redirect(routes.TravellerTypeController.updateDestinationTravellerType(destId));
     }
 }
